fix(falcon-server-env): Fixing default re-exports, aligning with other packages
