+++ packages/falcon-server-env/src/index.ts
@@ -1,9 +1,7 @@
 export { InMemoryLRUCache } from 'apollo-server-caching';
-export * from './models/ApiDataSource';
-export * from './models/EndpointManager';
-export { default as ContextHTTPCache } from './cache/ContextHTTPCache';
-export * from './cache/Cache';
-export * from './types';
-export { default as htmlHelpers } from './helpers/htmlHelpers';
-export { Events } from './events';
 export * from './authorization';
+export * from './cache';
+export * from './events';
+export * from './helpers';
+export * from './models';
+export * from './types';
